requirements.txt
streamlit
streamlit_cropper
python-dotenv
fpdf
Pillow
google-generativeai
langchain
tesserocr
